requirements/requirements_dev.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosmtplib==2.0.2
alembic==1.11.1
amqp==5.1.1
anyio==3.6.2
astor==0.8.1
async-timeout==4.0.2
attrs==23.1.0
bandit==1.7.5
bcrypt==4.0.1
billiard==4.1.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.2.5
cryptography==40.0.2
darglint==1.8.1
dnspython==2.3.0
docutils==0.20.1
ecdsa==0.18.0
email-validator==1.3.1
eradicate==2.3.0
factory-boy==3.2.1
Faker==18.9.0
fakeredis==2.15.0
fastapi==0.95.1
fastapi-cache2==0.2.1
fastapi-mail==1.2.8
fastapi-pagination==0.12.3
flake8==6.0.0
flake8-bandit==4.1.1
flake8-broken-line==1.0.0
flake8-bugbear==23.6.5
flake8-commas==2.1.0
flake8-comprehensions==3.13.0
flake8-debugger==4.1.2
flake8-docstrings==1.7.0
flake8-eradicate==1.5.0
flake8-isort==6.0.0
flake8-quotes==3.3.2
flake8-rst-docstrings==0.3.0
flake8-string-format==0.3.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
kombu==5.3.1
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.4
packaging==23.1
passlib==1.7.4
pbr==5.11.1
pendulum==2.1.2
pep8-naming==0.13.3
pluggy==1.0.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
pyasn1==0.5.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pydocstyle==6.3.0
pyflakes==3.0.1
Pygments==2.15.1
pytest==7.3.1
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytzdata==2020.1
PyYAML==6.0
redis==4.5.5
restructuredtext-lint==1.4.0
rich==13.4.2
rsa==4.9
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.13
SQLAlchemy-ImageAttach==1.1.0
SQLAlchemy-Utils==0.41.1
starlette==0.26.1
stevedore==5.1.0
toml==0.10.2
typing_extensions==4.5.0
tzdata==2023.3
uvicorn==0.22.0
vine==5.0.0
vulture==2.7
Wand==0.6.11
wcwidth==0.2.6
wemake-python-styleguide==0.18.0
yarl==1.9.2
